refactor: Move reference resolving logic to a separate module
